imeage_web_store/requirements.txt
asgiref==3.5.2
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==5.1.1
deprecation==2.1.0
Django==3.2.14
idna==3.3
img2pdf==0.4.4
importlib-metadata==4.12.0
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pikepdf==5.4.0
Pillow==9.2.0
prompt-toolkit==3.0.30
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
zipp==3.8.1
